+++ src/components/ui/toaster/types.ts
@@ -1,9 +1,0 @@
-interface Toast {
-    id: string;
-    text: string;
-    hide: boolean;
-}
-
-export type {
-    Toast,
-};